+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.11.0]
+- Remove storeCurrency product tracking
+- Add Analytics currency tracking init param
+- Add temporary file clean after store index
+- Add customer tracking using their customer id
+
 ## [4.10.0]
 
 - Add Tooso SDK configurations